feat: Add support for NRS spec and improve tests
This commit introduces support for the NRS specification by updating the Prisma schema to include the `Entry`, `EntryProgress`, and `EntryStatus` models. It also adds support for `DAH_additional_sources` by adding a JSON field to the `Entry` model.

The tests have been significantly improved to cover more scenarios, use more interesting data, and follow the NRS ID specification. The tests are now also able to be run in parallel.